refactor(agent): extract message factory
